cdk/pyproject.toml
[tool.poetry]
name = "cdk"
version = "1.0.0"
description = "AWS CDK Stack CDK Goat"
authors = ["Avishay Bar  <avishay.il@gmail.com>"]
packages = [{include = "cdk_goat_service"}]

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = ">=2.96.2"
cdk_ecr_deployment = ">=2.5.6"
constructs = ">=10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"